LayoutTests/fast/scrolling/progammatic-scroll-scrolling-no-frame-expected.txt


PASS iframe.contentWindow.scrollY is 0
PASS iframe.contentWindow.scrollY is 400
PASS successfullyParsed is true

TEST COMPLETE
